Remove abstract methods from SearchViewModelBase. Refactor setList method inside SearchViewModelBase to be a generic and not just play List
