+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/patches/PatchProposalPanel.java
@@ -106,7 +106,7 @@ public class PatchProposalPanel {
     protected JComponent createReturnToListSideComponent() {
         return ReturnToListComponent.INSTANCE.createReturnToListSideComponent(RadicleBundle.message("backToList"),
                 () -> {
-                     controller.createPanel();
+                    controller.createPanel();
                     return Unit.INSTANCE;
                 });
     }
